+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## 1.18.3 (TBD)
 
+- Added mastery for Starmageddon
+- Made Starmageddon projectile target random spots in the current room instead of the player when no enemies are found
 - Fixed null deref caused by improperly updating Zag projectile trails upon colliding with walls
 
 ## 1.18.2 (2024-09-03)